+++ feature/mypage/src/main/java/com/record/mypage/follow/FollowingScreen.kt
@@ -21,7 +21,7 @@ fun FollowingRoute(
     padding: PaddingValues,
     modifier: Modifier = Modifier,
     viewModel: FollowViewModel = hiltViewModel(),
-    navigateToProfile: (Int) -> Unit,
+    navigateToProfile: (Long) -> Unit,
 ) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
